Add files via upload
For this project, multiple sequence alignment and pairwise sequence alignment were performed and a heatmap was created from the scoring matrix of the pairwise sequence alignment of and isolated recombinant scFv protein, gene 16. A multiple sequence alignment was first performed as a means to visualize the alignment. Then a pairwise sequence alignment was performed and a heatmap was generated from the similarities of the sequences from the pairwise alignment. The purpose of the pairwise and the heatmap are to visualize the sequence alignment with more detail to answer the scientific question being asked, is gene_16 present and are there mutations?